test: cover toCsv special characters
