+++ src/app/api/export/route.ts
@@ -1,27 +1,12 @@
 import { NextResponse } from "next/server";
 import { createClient } from "@supabase/supabase-js";
 import { getCurrentUserId } from "@/lib/auth";
+import { toCsv } from "@/lib/csv";
 
 const supabase = createClient(
   process.env.NEXT_PUBLIC_SUPABASE_URL!,
   process.env.SUPABASE_SERVICE_ROLE_KEY!
 );
-
-function toCsv<T extends Record<string, unknown>>(rows: T[]): string {
-  if (!rows.length) return "";
-  const headers = Object.keys(rows[0]);
-  const escape = (val: unknown) => {
-    if (val === null || val === undefined) return "";
-    return `"${String(val).replace(/"/g, '""')}"`;
-  };
-  const lines = [
-    headers.join(","),
-    ...rows.map((row) =>
-      headers.map((h) => escape(row[h as keyof T])).join(","),
-    ),
-  ];
-  return lines.join("\n");
-}
 
 export async function GET(req: Request) {
   try {
