+++ src/lib/csv.ts
@@ -1,0 +1,15 @@
+export function toCsv<T extends Record<string, unknown>>(rows: T[]): string {
+  if (!rows.length) return "";
+  const headers = Object.keys(rows[0]);
+  const escape = (val: unknown) => {
+    if (val === null || val === undefined) return "";
+    return `"${String(val).replace(/"/g, '""')}"`;
+  };
+  const lines = [
+    headers.join(","),
+    ...rows.map((row) =>
+      headers.map((h) => escape(row[h as keyof T])).join(","),
+    ),
+  ];
+  return lines.join("\n");
+}
